updated the scraper with a new parser
