--- scraper/static/index.html
@@ -1,7 +1,0 @@
-<html>
-  <h1>South African COVID-19 Data Update Service</h1>
-  <a href="/get-updated-csv">Download updated data</a>
-  <hr>
-  <a href="https://github.com/dsfsi/covid19za">Data</a>
-  <p>This service is maintained by <a href="https://github.com/cishiv">Shivan Moodley</a></p>
-</html>
